Add Java specific makefile and update readme
The readme update reflects the lack of GNU specificity, and the addition of restriction descriptions within the file.

The makefile has many options and was written in a flexible manner.
Instructions are provided for changing parts of the functionality.
The makefile detects new source files recursively in (by default) src/, and creates the corresponding structure the compiler expects in bin/.
It also updates the classpath to include all compiled files.
When running, a useful error message is given to the user if they do not provide the location of the main class.
Modification of the running directive and main class variable name by the user are also supported.